feat: Enhance Dashboard layout with responsive design improvements
- Introduced a new mainLinks array for navigation, improving link management.
- Enhanced header section with responsive adjustments, including updated text sizes and gradient styling.
- Improved profile display with better image sizing and additional user information.
- Added a quote section for daily inspiration, enhancing visual appeal and responsiveness.
- Refined navigation layout for better interaction and responsiveness across devices.
